+++ lprint-modify.c
@@ -25,28 +25,35 @@ lprintDoModify(
 {
   http_t	*http;			// Connection to server
   ipp_t		*request;		// Create-Printer request
-  const char	*printer_name;		// Name of printer
+  const char	*printer_uri,		// Printer URI
+		*printer_name;		// Name of printer
   char		resource[1024];		// Resource path
 
 
-  // Get required values...
-  printer_name  = cupsGetOption("printer-name", num_options, options);
-
-  if (!printer_name)
+  // Open a connection to the server...
+  if ((printer_uri = cupsGetOption("printer-uri", num_options, options)) != NULL)
   {
-    if (!printer_name)
-      fputs("lprint: Missing '-d PRINTER' option.\n", stderr);
+    // Connect to the remote server...
+    if ((http = lprintConnectURI(printer_uri, resource, sizeof(resource))) == NULL)
+      return (1);
+  }
+  else if ((http = lprintConnect(1)) == NULL)
+  {
+    return (1);
+  }
+  else if ((printer_name  = cupsGetOption("printer-name", num_options, options)) == NULL)
+  {
+    fputs("lprint: Missing '-d PRINTER' option.\n", stderr);
 
     return (1);
   }
 
-  // Open a connection to the server...
-  if ((http = lprintConnect(1)) == NULL)
-    return (1);
-
   // Send a Set-Printer-Attributes request to the server...
   request = ippNewRequest(IPP_OP_SET_PRINTER_ATTRIBUTES);
-  lprintAddPrinterURI(request, printer_name, resource, sizeof(resource));
+  if (printer_uri)
+    ippAddString(request, IPP_TAG_PRINTER, IPP_TAG_URI, "printer-uri", NULL, printer_uri);
+  else
+    lprintAddPrinterURI(request, printer_name, resource, sizeof(resource));
   lprintAddOptions(request, num_options, options);
 
   ippDelete(cupsDoRequest(http, request, resource));