+++ CHANGES.md
@@ -5,6 +5,8 @@ v1.0b2 - January ??, 2020
 -------------------------
 
 - Added support for authentication of remote administrative requests (Issue #1)
+- The `add`, `default`, `delete`, and `modify` sub-commands now support the "-u"
+  option (Issue #1)
 - Fixed an ASLR issue with some Linux compilers (Issue #5)
 - Added code to detach the Linux usblp kernel driver since apparently there are
   still Linux distributions shipping the old (and broken) usblp kernel driver